matt: WIP created a winner page and filled out the home page
