+++ src/pages/HomePage.jsx
@@ -1,0 +1,87 @@
+import React from 'react'
+import { useEffect, useState } from 'react'
+import { useGetPokemon } from '../hooks/useGetPokemon'
+import { useGlobalPokemonData } from '../context/globalPokemonList'
+import { useGlobalPlayerData } from '../context/globalPlayerData'
+import { useStorage } from '../hooks/useStorage'
+import PokemonPickerList from '../components/PokemonPickerList'
+import PokemonCard from '../components/PokemonCard'
+import LoadingPokeball from '../components/LoadingPokeball'
+import { useNavigate } from 'react-router-dom'
+
+
+export default function HomePage() {
+    const { getPokemon, isLoading, error } = useGetPokemon()
+    const { pokemonList, setPokemonList } = useGlobalPokemonData()
+    const { setPersistenPokemonList } = useStorage()
+    const { playerData, playerDispatch } = useGlobalPlayerData()
+    const [activePlayer, setActivePlayer] = useState(1)
+    const navigate = useNavigate()
+
+    // Make all the API calls and get pokemon data on page refresh
+    useEffect(() => {
+        getPokemon()
+    }, [])
+
+    useEffect(() => {
+        // console.log(pokemonList[20])
+    }, [pokemonList])
+
+    // Button handler to change which player is choosing a pokemon
+    const handleChangeActivePlayer = () => {
+        setActivePlayer(activePlayer === 1 ? 2 : 1)
+        console.log(activePlayer)
+    }
+
+    const startBattle = () => {
+        navigate('/battle')
+    }
+
+
+    return (
+        <div>
+            {isLoading ?
+                (
+                    <LoadingPokeball />
+                ) :
+                (
+                    <div className='max-h-screen grid grid-rows-[auto_3fr_2fr]'>
+                        <h1 className='text-5xl text-center p-2'>
+                            Pokemon Battler
+                        </h1>
+
+                        <button
+                            onClick={handleChangeActivePlayer}
+                            className='border-2 border-pink-500 bg-pink-500/10 px-4 py-2 rounded-lg hover:bg-pink-500 hover:text-white active:bg-pink-400 md:mx-auto'
+                        >
+                            Player {activePlayer} Select
+                        </button>
+
+                        <button
+                            onClick={startBattle}
+                            className='border-2 border-pink-500 bg-pink-500/10 px-4 py-2 rounded-lg hover:bg-pink-500 hover:text-white active:bg-pink-400 md:mx-auto'
+                        >
+                            Battle!
+                        </button>
+
+                        <div className='grid grid-cols-[1fr_auto_1fr] gap-3 px-2'>
+                            <PokemonCard
+                                playerNum={1}
+                                pokemon={playerData.player1}
+                            />
+                            <span className='self-center text-5xl'>
+                                VS
+                            </span>
+                            <PokemonCard
+                                playerNum={2}
+                                pokemon={playerData.player2}
+                            />
+                        </div>
+
+                        <PokemonPickerList playerNum={activePlayer} />
+                    </div>
+                )
+            }
+        </div>
+    )
+}
