+++ src/pages/WinPage.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function WinPage() {
+    return (
+        <div>Winner!</div>
+    )
+}
